pyproject.toml
[build-system]
requires = ["setuptools>=80.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "civilib"
version = "0.1.0"
description = "Shared code for civiliza project"
requires-python = ">=3.8"
dependencies = [
    "pydantic[email]",
    "cognitojwt[sync]",
    "python-ulid",
    "email-validator",
    "twilio",
    "cefapi @ git+ssh://git@github.com/lucasjrt/cefapi@main"
]

[tool.setuptools.packages.find]
include = ["*"]

[tool.pytest_env]
DYNAMODB_TABLE = "civilib-test"
AWS_ACCESS_KEY_ID = "testing"
AWS_SECRET_ACCESS_KEY = "testing"
AWS_SESSION_TOKEN = "testing"
AWS_DEFAULT_REGION = "sa-east-1"
STORAGE_BUCKET = "incc-app-storage-8ff2cd83"
